app/utils/ehospital_service.py
# Secure, robust ehospital_service.py implementation
import requests
import logging
from marshmallow import Schema, fields, ValidationError
from tenacity import retry, stop_after_attempt, wait_fixed

class EHospitalResponseSchema(Schema):
	status = fields.String(required=True)
	data = fields.Dict(required=False)

def validate_params(params):
	# Example: Validate and sanitize input parameters
	if not isinstance(params, dict):
		raise ValueError('Parameters must be a dictionary')
	for k, v in params.items():
		if not isinstance(k, str) or not isinstance(v, (str, int, float)):
			raise ValueError('Invalid parameter type')
	return params

@retry(stop=stop_after_attempt(3), wait=wait_fixed(2))
def call_ehospital_api(url, params, timeout=10):
	validate_params(params)
	try:
		response = requests.post(url, json=params, timeout=timeout)
		response.raise_for_status()
		resp_json = response.json()
		# Validate response schema
		schema = EHospitalResponseSchema()
		schema.load(resp_json)
		return resp_json
	except ValidationError as ve:
		logging.error(f"Response schema validation error: {ve}")
		raise
	except requests.RequestException as re:
		logging.error(f"Request error: {re}")
		raise
