requirements.txt
flask==3.1.2
python-dotenv==1.1.1
flask-restx==1.3.0
flask-cors==6.0.1
flask-limiter==3.12
marshmallow==4.0.1
celery==5.5.3
redis==6.4.0
requests==2.32.3
pytest==8.4.1
pytest-flask==1.3.0
flask-sqlalchemy==3.1.1
alembic==1.14.0
prometheus-client==0.21.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-instrumentation-flask==0.57b0
opentelemetry-instrumentation-requests==0.57b0
phonenumbers==8.13.50
